Refactor reusable code to separate folder
