+++ src/schema/IId.ts
@@ -1,0 +1,3 @@
+export interface IId {
+    id: string
+}
